Move `zero_for` to the `reference_executor_test` modules and remove duplicate test function.
PiperOrigin-RevId: 259595764
